checkpoint

+++ minio/build.sh
@@ -3,12 +3,13 @@ bash build.sh
 cd -
 cd ../ca-maker/
 bash build.sh
-sudo docker run -ti --rm -v minio-certs:/certs/server -v minio-ca:/certs/ca -e HOST=localhost -e GEN_NAME_KEY=private -e GEN_NAME_CRT=public ca:local
+[ -n "$(sudo docker volume ls -q -f name=minio-ca)" ] || sudo docker run -ti --rm -v minio-certs:/certs/server -v minio-ca:/certs/ca -e HOST=localhost -e GEN_NAME_KEY=private -e GEN_NAME_CRT=public ca:local
 cd -
-sudo docker-compose down -v
-sudo docker-compose up -d
-
-sudo docker run -ti --rm -v minio-ca:/certs alpine:latest cat /certs/ca.crt > /tmp/ca.crt
+[ -n "$(sudo docker network ls -q -f name=minio-network)" ] || sudo docker network create --driver bridge --attachable minio-network
+[ -n "$(sudo docker container ls -q -f name=minio)" ] || sudo docker-compose up -d
 
 # Necessary to run Minio python package locally with self signed certificates
+
+[ -n "$(sudo docker volume ls -q -f name=minio-ca)" ] || sudo docker run -ti --rm -v minio-ca:/certs alpine:latest cat /certs/ca.crt > /tmp/ca.crt
+
 export SSL_CERT_FILE=/tmp/ca.crt
