+++ postgres/build.sh
@@ -1,6 +1,9 @@
+cd ../traefik
+bash build.sh
+cd -
 cd ../ca-maker/
 bash build.sh
-sudo docker run -ti --rm -v postgres-certs:/certs/server -e OWNER=1001 -e GEN_NAME=postgres ca:local
+[ -n "$(sudo docker volume ls -q -f name=postgres-certs)" ] || sudo docker run -ti --rm -v postgres-certs:/certs/server -e OWNER=1001 -e GEN_NAME=postgres ca:local
 cd -
-sudo docker-compose down -v
-sudo docker-compose up -d
+[ -n "$(sudo docker network ls -q -f name=postgres-network)" ] || sudo docker network create --driver bridge --attachable postgres-network
+[ -n "$(sudo docker container ls -q -f name=postgres)" ] || sudo docker-compose up -d
